Most basic JWT authentication example
Showcase the minimal setup required to make a spring security JWT based
authentication example work.

We only include the resource server (oauth) speing boot starter, the
spring boot security starter and the spring web starter, along with the
public key location as configuration. With this setup, when sending
requests with the JWT also included within the classpath resources we
get the JWT based authentication to work.